refactor auth endpoint
